feat: 제품 추가 기능 구현 2022.09.07
제품 추가시 이미지를 cloudinary에서 관리한다.

+++ pages/api/products/[id].js
@@ -5,7 +5,10 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
+
+  const token = cookies.token;
 
   dbConnect();
 
@@ -19,6 +22,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "PUT") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("인증되지 않았습니다.");
+    }
     try {
       const product = await Product.create(req.body);
       res.status(201).json(product);
@@ -28,6 +34,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "DELETE") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("인증되지 않았습니다.");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("상품이 삭제되었습니다.");
